src/backend/requirements.txt
Flask == 1.1.1
Flask-REST == 1.3
Flask-RESTful == 0.3.7
Flask-Script == 2.0.6
Flask-SQLAlchemy == 2.4.1
Flask-JWT == 0.3.2
flask-mongoengine == 0.9.5
flask-marshmallow == 0.10.1
Flask-Injector == 0.12.3
marshmallow == 3.3.0
SQLAlchemy == 1.3.11
marshmallow-sqlalchemy == 0.21.0
pymongo == 3.10.0
passlib == 1.7.2
psycopg2 == 2.8.4
pandas==1.0.2
pdfkit==0.6.1
requests == 2.23.0
matplotlib == 3.2.1
numpy == 1.18.2
ptvsd==4.3.2
beautifulsoup4==4.9.3
#-e git+git@github.com:aistairc/qai-testbed.git#egg=ReportGenerator&subdirectory=src/backend/report
# install ReportGenerator
./report

